rename Product FilterRequest to IndexRequest
