+++ app/Http/Controllers/Admin/Product/IndexController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Admin\Product;
 
 use App\Http\Controllers\Controller;
 use App\Http\Filters\Product\ProductFilter;
-use App\Http\Requests\Admin\Product\FilterRequest;
+use App\Http\Requests\Admin\Product\IndexRequest;
 use App\Http\Requests\Admin\Product\IndexRequest;
 use App\Http\Resources\Admin\Product\ProductCollectionResource;
 use App\Models\Product;
@@ -12,7 +12,7 @@ use App\Models\Product;
 class IndexController extends Controller
 {
 
-    public function __invoke(FilterRequest $filterRequest)
+    public function __invoke(IndexRequest $filterRequest)
     {
         $data = $filterRequest->validated();
