+++ app/Http/Filters/Order/OrderSort.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Filters\Product;
+namespace App\Http\Filters\Order;
 
 use App\Http\Filters\AbstractFilter;
 use Illuminate\Database\Eloquent\Builder;
